refactor(webapp/rewards): move helpers to util
